Fix shifting in GetColorIndex
When shift is small, the result of (shift-1) and (shift-2) is negative. The intended result is to have a shift left instead of a shift right, but C# doesn't do it that way -- the expression results in 0.

Without this fix, colors will get binned (via AddColor) into the wrong (but nearby) leaf nodes in the Octree. The algorithm still works but may not generate the right number of colors (e.g. 64 instead of 256 in some cases), and error will be higher.

In particular, an image of a red->black gradient comprised of 256 colors will generate a palette of 64 colors because you only get 64 leaf nodes when binning those colors (without this fix).

This can significantly increase the memory consumption of the Octree -- by up to 4x.